# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/block": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Block"
                ],
                "summary": "Get all data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "default asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order By",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "keyword nama",
                        "name": "filter[name]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Block"
                ],
                "summary": "Create data",
                "parameters": [
                    {
                        "type": "string",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "row",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "slot",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "tier",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "yard",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            }
        },
        "/block/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Block"
                ],
                "summary": "Get spesifik data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Block"
                ],
                "summary": "Update data",
                "parameters": [
                    {
                        "type": "string",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "row",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "slot",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "tier",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "yard",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            }
        },
        "/pickup": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stacking"
                ],
                "summary": "Suggest Placement",
                "parameters": [
                    {
                        "type": "number",
                        "name": "container_height",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "container_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "yard",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            }
        },
        "/placement": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stacking"
                ],
                "summary": "Suggest Placement",
                "parameters": [
                    {
                        "type": "number",
                        "name": "container_height",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "container_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "yard",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            }
        },
        "/suggestion": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stacking"
                ],
                "summary": "Suggest Placement",
                "parameters": [
                    {
                        "type": "number",
                        "name": "container_height",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "container_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "yard",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.ReturnObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.ReturnObject": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}